rewrite ClientInner::on_datachannel_open()
This change removes ferrying events around between coordinator,
connection, etc. since it is unclear to me whether it brings any benefit
over doing this only in ClientInner::on_message().

This change also adds calls to task.substream_read_write() for the new
channel and sends the data added in that call to the remote. This is
required because smoldot sends a multistream-select header immediately
after an inbound substream is opened. If this is not flushed right away,
the header and protocol confirmation/rejection will be sent in a single
SCTP message that contains two libp2p-webrtc protobuf frames, which
litep2p does not handle correctly.